+++ src/GitHub_Updater/Base.php
@@ -16,6 +16,7 @@ use Fragen\Singleton,
 	Fragen\GitHub_Updater\API\Bitbucket_API,
 	Fragen\GitHub_Updater\API\Bitbucket_Server_API,
 	Fragen\GitHub_Updater\API\GitLab_API,
+	Fragen\GitHub_Updater\API\Gitea_API,
 	Fragen\GitHub_Updater\API\Language_Pack_API;
 
 
@@ -121,6 +122,8 @@ class Base {
 		'gitlab'            => false,
 		'gitlab_private'    => false,
 		'gitlab_enterprise' => false,
+		'gitea'             => false,
+		'gitea_private'     => false,
 	);
 
 	/**
@@ -148,6 +151,12 @@ class Base {
 			self::$git_servers['gitlab']        = 'GitLab';
 		} else {
 			self::$installed_apis['gitlab_api'] = false;
+		}
+		if ( class_exists( 'Fragen\GitHub_Updater\API\Gitea_API' ) ) {
+			self::$installed_apis['gitea_api'] = true;
+			self::$git_servers['gitea']        = 'Gitea';
+		} else {
+			self::$installed_apis['gitea_api'] = false;
 		}
 	}
 
@@ -476,6 +485,12 @@ class Base {
 			case 'gitlab_theme':
 				if ( self::$installed_apis['gitlab_api'] ) {
 					$this->repo_api = new GitLab_API( $repo );
+				}
+				break;
+			case 'gitea_plugin':
+			case 'gitea_theme':
+				if ( self::$installed_apis['gitea_api'] ) {
+					$this->repo_api = new Gitea_API( $repo );
 				}
 				break;
 		}
@@ -856,11 +871,13 @@ class Base {
 			'GitHub'    => 'github_' . $type,
 			'Bitbucket' => 'bitbucket_' . $type,
 			'GitLab'    => 'gitlab_' . $type,
+			'Gitea'     => 'gitea_' . $type,
 		);
 		$repo_base_uris = array(
 			'GitHub'    => 'https://github.com/',
 			'Bitbucket' => 'https://bitbucket.org/',
 			'GitLab'    => 'https://gitlab.com/',
+			'Gitea'     => '',
 		);
 
 		if ( array_key_exists( $repo, $repo_types ) ) {
@@ -1090,6 +1107,10 @@ class Base {
 			case 'gitlab_theme':
 				$this->repo_api = new GitLab_API( $repo );
 				break;
+			case 'gitea_plugin':
+			case 'gitea_theme':
+				$this->repo_api = new Gitea_API( $repo );
+				break;
 		}
 
 		$this->tag = isset( $_GET['rollback'] ) ? $_GET['rollback'] : null;
@@ -1232,6 +1253,37 @@ class Base {
 		$header['release_asset'] = ! $header['release_asset'] ? 'true' === $headers['Release Asset'] : $header['release_asset'];
 
 		return $header;
+	}
+
+	/**
+	 * Return an array of the running git servers.
+	 *
+	 * @access public
+	 * @return array $gits
+	 */
+	public function get_running_git_servers() {
+		$plugins = Singleton::get_instance( 'Plugin', $this )->get_plugin_configs();
+		$themes  = Singleton::get_instance( 'Theme', $this )->get_theme_configs();
+
+		$repos = array_merge( $plugins, $themes );
+		$gits  = array_map( function( $e ) {
+			if ( ! empty( $e->enterprise ) && false !== stripos( $e->type, 'bitbucket' ) ) {
+				return 'bbserver';
+			}
+
+			return $e->type;
+		}, $repos );
+
+		$gits = array_unique( array_values( $gits ) );
+
+		$gits = array_map( function( $e ) {
+			$e = explode( '_', $e );
+
+			return $e[0];
+		}, $gits );
+
+
+		return array_unique( $gits );
 	}
 
 	/**