+++ src/GitHub_Updater/Plugin.php
@@ -425,7 +425,8 @@ class Plugin extends Base {
 				);
 
 				// Skip on RESTful updating.
-				if ( isset( $_GET['action'] ) && 'github-updater-update' === $_GET['action'] &&
+				if ( isset( $_GET['action'], $_GET['plugin'] ) &&
+				     'github-updater-update' === $_GET['action'] &&
 				     $response['slug'] === $_GET['plugin']
 				) {
 					continue;